fix: show supplier name on the dashboard widget, reject fractional returns
Owner: "ويكتبلي في الداش بورد جمب الطلب هيتجاب من عند (المورد)" — the
"محتاجين نجيبها من المورد" widget resolves each stage instance's already-
tracked assignedSupplierId to a name and shows it next to the order.

Owner: "مينفعش أي كمية تكون بتقبل تبقى كسر" — the return-quantity input
accepted decimals (step="1" is only a spinner hint, not validation).
Added an explicit integer check client-side and .int() server-side on
createOrderItemReturnSchema, matching the original sale quantity's own
integer constraint.

+++ packages/shared/src/schemas/orderItemReturn.ts
@@ -18,7 +18,10 @@ export const orderItemReturnSchema = z.object({
 });
 
 export const createOrderItemReturnSchema = z.object({
-  quantity: z.number().positive(),
+  // Owner (2026-08-23, "مينفعش أي كمية تكون بتقبل تبقى كسر") — matches
+  // `inventoryRetailPricingInputSchema.quantity`'s own `.int()` (the
+  // original sale is always whole pieces, so a return of it must be too).
+  quantity: z.number().int().positive(),
   reason: z.string().trim().min(1).max(1000).optional(),
 });
